+++ apps/demo-react/providers/web3.tsx
@@ -1,18 +1,24 @@
 import { FC, PropsWithChildren, useMemo } from 'react';
-import {
-  AutoConnect,
-  ReefKnot,
-  getWalletsDataList,
-} from 'reef-knot/core-react';
+import { useThemeToggle } from '@lidofinance/lido-ui';
+import { ReefKnotProvider, getDefaultConfig } from 'reef-knot/core-react';
 import { WalletsListEthereum } from 'reef-knot/wallets';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { WagmiProvider, createConfig, http } from 'wagmi';
+import { WagmiProvider, http } from 'wagmi';
 import * as wagmiChains from 'wagmi/chains';
 import { CHAINS } from '@lido-sdk/constants';
+import type { Transport } from 'viem';
+import {
+  ReefKnotWalletsModal,
+  getDefaultWalletsModalConfig,
+} from 'reef-knot/connect-wallet-modal';
 
+import metrics from 'utils/metrics';
 import { getBackendRPCPath } from 'config';
 import { useClientConfig } from 'providers/client-config';
 import { SDKLegacyProvider } from './sdk-legacy';
+
+const LINK_DONT_HAVE_WALLET =
+  'https://support.metamask.io/hc/en-us/articles/360015489531-Getting-started-with-MetaMask';
 
 type ChainsList = [wagmiChains.Chain, ...wagmiChains.Chain[]];
 
@@ -27,6 +33,7 @@ const Web3Provider: FC<PropsWithChildren> = ({ children }) => {
     walletconnectProjectId,
     isWalletConnectionAllowed,
   } = useClientConfig();
+  const { themeName } = useThemeToggle();
 
   const { supportedChains, defaultChain } = useMemo(() => {
     const supportedChains = wagmiChainsArray.filter((chain) =>
@@ -55,40 +62,54 @@ const Web3Provider: FC<PropsWithChildren> = ({ children }) => {
     [supportedChainIds],
   );
 
-  const { walletsDataList } = useMemo(() => {
-    return getWalletsDataList({
-      walletsList: WalletsListEthereum,
-      rpc: backendRPC,
-      walletconnectProjectId: walletconnectProjectId,
-      defaultChain: defaultChain,
-    });
-  }, [backendRPC, defaultChain, walletconnectProjectId]);
-
-  const config = useMemo(() => {
-    return createConfig({
-      chains: supportedChains,
-      ssr: true,
-      multiInjectedProviderDiscovery: false,
-      transports: supportedChains.reduce(
-        (res, curr) => ({
-          ...res,
-          [curr.id]: http(backendRPC[curr.id], { batch: true }),
-        }),
-        {},
-      ),
-    });
+  const transports = useMemo(() => {
+    return supportedChains.reduce<Record<number, Transport>>(
+      (res, curr) => ({
+        ...res,
+        [curr.id]: http(backendRPC[curr.id], { batch: true }),
+      }),
+      {},
+    );
   }, [supportedChains, backendRPC]);
+
+  const { wagmiConfig, reefKnotConfig, walletsModalConfig } = useMemo(() => {
+    return getDefaultConfig({
+      // Reef-Knot config args
+      rpc: backendRPC,
+      defaultChain: defaultChain,
+      walletconnectProjectId,
+      walletsList: WalletsListEthereum,
+
+      // Wagmi config args
+      transports,
+      chains: supportedChains,
+      autoConnect: isWalletConnectionAllowed,
+      ssr: true,
+
+      // Wallets config args
+      // TODO: We could call `getDefaultWalletsModalConfig` inside `getDefaultConfig`, but it cause package dependency cycle rn
+      ...getDefaultWalletsModalConfig(),
+      metrics,
+      linkDontHaveWallet: LINK_DONT_HAVE_WALLET,
+    });
+  }, [
+    backendRPC,
+    supportedChains,
+    defaultChain,
+    walletconnectProjectId,
+    isWalletConnectionAllowed,
+    transports,
+  ]);
 
   return (
     // default wagmi autoConnect, MUST be false in our case, because we use custom autoConnect from Reef Knot
-    <WagmiProvider config={config} reconnectOnMount={false}>
+    <WagmiProvider config={wagmiConfig} reconnectOnMount={false}>
       <QueryClientProvider client={queryClient}>
-        <ReefKnot
-          rpc={backendRPC}
-          chains={supportedChains}
-          walletDataList={walletsDataList}
-        >
-          {isWalletConnectionAllowed && <AutoConnect autoConnect />}
+        <ReefKnotProvider config={reefKnotConfig}>
+          <ReefKnotWalletsModal
+            config={walletsModalConfig}
+            darkThemeEnabled={themeName === 'dark'}
+          />
           <SDKLegacyProvider
             defaultChainId={defaultChain.id}
             supportedChains={supportedChains}
@@ -96,7 +117,7 @@ const Web3Provider: FC<PropsWithChildren> = ({ children }) => {
           >
             {children}
           </SDKLegacyProvider>
-        </ReefKnot>
+        </ReefKnotProvider>
       </QueryClientProvider>
     </WagmiProvider>
   );