+++ packages/connect-wallet-modal/src/components/ConnectButtonBase/ConnectButtonBase.tsx
@@ -1,5 +1,4 @@
 import React, { ElementType, FC, isValidElement } from 'react';
-import { WalletAdapterData, WalletAdapterIcons } from '@reef-knot/types';
 import {
   ConnectButtonStyle,
   ConnectButtonContentStyle,
@@ -7,23 +6,13 @@ import {
   ConnectButtonIconStyle,
   ConnectButtonTitleStyle,
 } from './styles';
-import { ConnectButtonProps } from '../../connectButtons/types';
+import { ConnectButtonBaseProps } from '@reef-knot/types';
+import { isWalletAdapterIcons } from '../../helpers/isWalletAdapterIcons';
 
-// Check that icon's type is WalletAdapterIcons by looking at its "light" and "dark" fields
-function isWalletAdapterIcons(
-  icon: WalletAdapterData['icon'],
-): icon is WalletAdapterIcons {
-  return Boolean(
-    icon &&
-      (icon as WalletAdapterIcons).light &&
-      (icon as WalletAdapterIcons).dark,
-  );
-}
-
-const ConnectButton: FC<ConnectButtonProps> = (props: ConnectButtonProps) => {
+export const ConnectButtonBase: FC<ConnectButtonBaseProps> = (props) => {
   const {
     icon,
-    shouldInvertWalletIcon,
+    darkThemeEnabled,
     children,
     isCompact,
     isLoading = false,
@@ -33,7 +22,7 @@ const ConnectButton: FC<ConnectButtonProps> = (props: ConnectButtonProps) => {
   let ButtonIcon: ElementType = React.Fragment;
   if (icon) {
     if (isWalletAdapterIcons(icon)) {
-      ButtonIcon = shouldInvertWalletIcon ? icon.dark : icon.light;
+      ButtonIcon = darkThemeEnabled ? icon.dark : icon.light;
     } else {
       ButtonIcon = icon;
     }
@@ -52,5 +41,3 @@ const ConnectButton: FC<ConnectButtonProps> = (props: ConnectButtonProps) => {
     </ConnectButtonStyle>
   );
 };
-
-export default ConnectButton;