+++ packages/connect-wallet-modal/src/components/WalletsModal/index.ts
@@ -1,3 +1,0 @@
-export * from './WalletsModal';
-export * from './WalletsModalForEth';
-export * from './types';