+++ packages/connect-wallet-modal/src/connectButtons/ConnectInjected.tsx
@@ -2,8 +2,9 @@ import React, { FC, useCallback } from 'react';
 import { useConnect } from 'wagmi';
 import { useDisconnect } from '@reef-knot/core-react';
 import { isMobileOrTablet } from '@reef-knot/wallets-helpers';
-import { ConnectButton } from '../components/ConnectButton';
-import { suggestApp, openWindow } from '../helpers';
+import { ConnectButtonBase } from '../components/ConnectButtonBase';
+import { suggestApp } from '../helpers/suggestApp';
+import { openWindow } from '../helpers/openWindow';
 import { ConnectInjectedProps } from './types';
 
 export const ConnectInjected: FC<ConnectInjectedProps> = (
@@ -11,8 +12,7 @@ export const ConnectInjected: FC<ConnectInjectedProps> = (
 ) => {
   const {
     onConnect,
-    onBeforeConnect,
-    shouldInvertWalletIcon,
+    darkThemeEnabled,
     metrics,
     walletId,
     walletName,
@@ -30,7 +30,6 @@ export const ConnectInjected: FC<ConnectInjectedProps> = (
   const { disconnect } = useDisconnect();
 
   const handleConnect = useCallback(async () => {
-    onBeforeConnect?.();
     metricsOnClick?.();
 
     if (await detector?.()) {
@@ -51,21 +50,20 @@ export const ConnectInjected: FC<ConnectInjectedProps> = (
     disconnect,
     downloadURLs,
     metricsOnClick,
-    onBeforeConnect,
     onConnect,
     metricsOnConnect,
   ]);
 
   return (
-    <ConnectButton
+    <ConnectButtonBase
       {...rest}
       icon={WalletIcon}
-      shouldInvertWalletIcon={shouldInvertWalletIcon}
+      darkThemeEnabled={darkThemeEnabled}
       onClick={() => {
         void handleConnect();
       }}
     >
       {walletName}
-    </ConnectButton>
+    </ConnectButtonBase>
   );
 };