+++ packages/connect-wallet-modal/src/connectButtons/types.ts
@@ -1,16 +1,5 @@
-import { ButtonProps } from '@reef-knot/ui-react';
-import { WalletAdapterData } from '@reef-knot/types';
-import { ButtonsCommonProps } from '../components';
+import type { ConnectWalletButtonProps } from '@reef-knot/types';
 
-export type ConnectButtonProps = {
-  icon: WalletAdapterData['icon'];
-  shouldInvertWalletIcon?: boolean;
-  isCompact?: boolean;
-  isLoading?: boolean;
-} & ButtonProps;
-
-export type ConnectWalletProps = ButtonsCommonProps & ButtonProps;
-
-export type ConnectInjectedProps = WalletAdapterData & ConnectWalletProps;
-export type ConnectWCProps = WalletAdapterData & ConnectWalletProps;
-export type ConnectLedgerProps = WalletAdapterData & ConnectWalletProps;
+export type ConnectInjectedProps = ConnectWalletButtonProps;
+export type ConnectWCProps = ConnectWalletButtonProps;
+export type ConnectLedgerProps = ConnectWalletButtonProps;