+++ packages/connect-wallet-modal/src/helpers/sortWalletsList.ts
@@ -1,10 +1,12 @@
-import type { WalletAdapterData } from '@reef-knot/types';
-import type { WalletsModalProps } from '../../types';
+import type {
+  ReefKnotWalletsModalConfig,
+  WalletAdapterData,
+} from '@reef-knot/types';
 
 type GetWalletsListArgs = {
   walletDataList: WalletAdapterData[];
-  walletsShown: WalletsModalProps['walletsShown'];
-  walletsPinned: WalletsModalProps['walletsPinned'];
+  walletsShown: ReefKnotWalletsModalConfig['walletsShown'];
+  walletsPinned: ReefKnotWalletsModalConfig['walletsPinned'];
 };
 
 type FilteredWalletData = {