+++ packages/reef-knot/CHANGELOG.md
@@ -1,5 +1,20 @@
 # reef-knot
 
+## 6.0.0
+
+### Major Changes
+
+- reef-knot setup reworked
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/connect-wallet-modal@6.0.0
+  - @reef-knot/core-react@5.0.0
+  - @reef-knot/types@3.0.0
+  - @reef-knot/web3-react@5.0.0
+  - @reef-knot/wallets-list@3.0.0
+
 ## 5.7.6
 
 ### Patch Changes