+++ packages/wallets/binance-wallet/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @reef-knot/wallet-adapter-binance-wallet
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/types@3.0.0
+
 ## 1.0.3
 
 ### Patch Changes