+++ packages/wallets/exodus/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @reef-knot/wallet-adapter-exodus
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/types@3.0.0
+
 ## 2.1.1
 
 ### Patch Changes