+++ packages/web3-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @reef-knot/web3-react
 
+## 5.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/core-react@5.0.0
+
 ## 4.0.2
 
 ### Patch Changes